Address CodeRabbit env migration comments
