+++ src/core/config/parse/env_registry.rs
@@ -26,14 +26,6 @@ pub(crate) enum LegacyBehavior {
     Advanced,
 }
 
-use self::EnvClassification::{
-    CompatibilityShim, ComposeEnv, Delete, KeepEnv, MoveToml, TrustedOperatorBootstrap,
-};
-use self::LegacyBehavior::{
-    Advanced, Canonical, DeleteOnMigration, WarnAndIgnore, WarnEnvOverride,
-};
-use self::RuntimePlacement::{Both, ComposeInterpolation, ContainerRequired, HostOnly, NotRuntime};
-
 #[derive(Debug, Clone, Copy)]
 #[allow(dead_code)]
 pub(crate) struct EnvKeySpec {
@@ -45,755 +37,14 @@ pub(crate) struct EnvKeySpec {
     pub secret: bool,
 }
 
-pub(crate) const ENV_KEY_SPECS: &[EnvKeySpec] = &[
-    spec("QDRANT_URL", KeepEnv, Both, None, Canonical, false),
-    spec("TEI_URL", KeepEnv, Both, None, Canonical, false),
-    spec(
-        "AXON_CHROME_REMOTE_URL",
-        KeepEnv,
-        Both,
-        None,
-        Canonical,
-        false,
-    ),
-    spec("AXON_SERVER_URL", KeepEnv, HostOnly, None, Canonical, false),
-    spec(
-        "AXON_MCP_HTTP_TOKEN",
-        KeepEnv,
-        ContainerRequired,
-        None,
-        Canonical,
-        true,
-    ),
-    spec("AXON_MCP_AUTH_MODE", KeepEnv, Both, None, Canonical, false),
-    spec("AXON_MCP_PUBLIC_URL", KeepEnv, Both, None, Canonical, false),
-    spec(
-        "AXON_MCP_GOOGLE_CLIENT_ID",
-        KeepEnv,
-        Both,
-        None,
-        Canonical,
-        false,
-    ),
-    spec(
-        "AXON_MCP_GOOGLE_CLIENT_SECRET",
-        KeepEnv,
-        Both,
-        None,
-        Canonical,
-        true,
-    ),
-    spec(
-        "AXON_MCP_AUTH_ADMIN_EMAIL",
-        KeepEnv,
-        Both,
-        None,
-        Canonical,
-        false,
-    ),
-    spec(
-        "AXON_MCP_AUTH_ALLOWED_REDIRECT_URIS",
-        KeepEnv,
-        Both,
-        None,
-        Canonical,
-        false,
-    ),
-    spec(
-        "AXON_MCP_ALLOWED_ORIGINS",
-        KeepEnv,
-        Both,
-        None,
-        Canonical,
-        false,
-    ),
-    spec(
-        "AXON_ENV_FILE",
-        TrustedOperatorBootstrap,
-        HostOnly,
-        None,
-        Advanced,
-        false,
-    ),
-    spec(
-        "AXON_CONFIG_PATH",
-        TrustedOperatorBootstrap,
-        HostOnly,
-        None,
-        Advanced,
-        false,
-    ),
-    spec(
-        "AXON_HOME",
-        TrustedOperatorBootstrap,
-        Both,
-        None,
-        Advanced,
-        false,
-    ),
-    spec(
-        "AXON_DATA_DIR",
-        TrustedOperatorBootstrap,
-        Both,
-        None,
-        Advanced,
-        false,
-    ),
-    spec(
-        "AXON_SQLITE_PATH",
-        TrustedOperatorBootstrap,
-        Both,
-        None,
-        Advanced,
-        false,
-    ),
-    spec(
-        "AXON_MCP_HTTP_PUBLISH",
-        ComposeEnv,
-        ComposeInterpolation,
-        None,
-        Canonical,
-        false,
-    ),
-    spec(
-        "AXON_IMAGE",
-        ComposeEnv,
-        ComposeInterpolation,
-        None,
-        Canonical,
-        false,
-    ),
-    spec(
-        "GEMINI_HOME",
-        ComposeEnv,
-        ComposeInterpolation,
-        None,
-        Canonical,
-        false,
-    ),
-    spec(
-        "TEI_EMBEDDING_MODEL",
-        ComposeEnv,
-        ComposeInterpolation,
-        None,
-        Canonical,
-        false,
-    ),
-    spec(
-        "TEI_HTTP_PORT",
-        ComposeEnv,
-        ComposeInterpolation,
-        None,
-        Canonical,
-        false,
-    ),
-    spec(
-        "TEI_SERVER_MAX_CLIENT_BATCH_SIZE",
-        ComposeEnv,
-        ComposeInterpolation,
-        None,
-        Canonical,
-        false,
-    ),
-    spec(
-        "NVIDIA_VISIBLE_DEVICES",
-        ComposeEnv,
-        ComposeInterpolation,
-        None,
-        Canonical,
-        false,
-    ),
-    spec(
-        "CUDA_VISIBLE_DEVICES",
-        ComposeEnv,
-        ComposeInterpolation,
-        None,
-        Canonical,
-        false,
-    ),
-    spec("TAVILY_API_KEY", KeepEnv, Both, None, Canonical, true),
-    spec("GITHUB_TOKEN", KeepEnv, Both, None, Canonical, true),
-    spec("REDDIT_CLIENT_ID", KeepEnv, Both, None, Canonical, false),
-    spec("REDDIT_CLIENT_SECRET", KeepEnv, Both, None, Canonical, true),
-    spec(
-        "HF_TOKEN",
-        KeepEnv,
-        ComposeInterpolation,
-        None,
-        Canonical,
-        true,
-    ),
-    spec(
-        "OPENAI_MODEL",
-        CompatibilityShim,
-        Both,
-        None,
-        WarnEnvOverride,
-        false,
-    ),
-    spec(
-        "OPENAI_BASE_URL",
-        CompatibilityShim,
-        Both,
-        None,
-        WarnAndIgnore,
-        false,
-    ),
-    spec(
-        "OPENAI_API_KEY",
-        CompatibilityShim,
-        Both,
-        None,
-        WarnAndIgnore,
-        true,
-    ),
-    spec(
-        "AXON_HEADLESS_GEMINI_CMD",
-        KeepEnv,
-        Both,
-        None,
-        Canonical,
-        false,
-    ),
-    spec(
-        "AXON_HEADLESS_GEMINI_HOME",
-        TrustedOperatorBootstrap,
-        Both,
-        None,
-        Advanced,
-        false,
-    ),
-    spec(
-        "AXON_HEADLESS_GEMINI_MODEL",
-        KeepEnv,
-        Both,
-        None,
-        Canonical,
-        false,
-    ),
-    spec(
-        "AXON_LLM_COMPLETION_CONCURRENCY",
-        KeepEnv,
-        Both,
-        None,
-        Canonical,
-        false,
-    ),
-    spec(
-        "AXON_LLM_COMPLETION_TIMEOUT_SECS",
-        KeepEnv,
-        Both,
-        None,
-        Canonical,
-        false,
-    ),
-    spec(
-        "TEI_MAX_CLIENT_BATCH_SIZE",
-        MoveToml,
-        NotRuntime,
-        Some("tei.max-client-batch-size"),
-        WarnEnvOverride,
-        false,
-    ),
-    spec(
-        "TEI_MAX_RETRIES",
-        MoveToml,
-        NotRuntime,
-        Some("tei.max-retries"),
-        WarnEnvOverride,
-        false,
-    ),
-    spec(
-        "TEI_REQUEST_TIMEOUT_MS",
-        MoveToml,
-        NotRuntime,
-        Some("tei.request-timeout-ms"),
-        WarnEnvOverride,
-        false,
-    ),
-    spec(
-        "AXON_INGEST_LANES",
-        MoveToml,
-        NotRuntime,
-        Some("workers.ingest-lanes"),
-        WarnEnvOverride,
-        false,
-    ),
-    spec(
-        "AXON_EMBED_LANES",
-        MoveToml,
-        NotRuntime,
-        Some("workers.embed-lanes"),
-        WarnEnvOverride,
-        false,
-    ),
-    spec(
-        "AXON_EMBED_DOC_TIMEOUT_SECS",
-        MoveToml,
-        NotRuntime,
-        Some("workers.embed-doc-timeout-secs"),
-        WarnEnvOverride,
-        false,
-    ),
-    spec(
-        "AXON_QUEUE_SUMMARY_SECS",
-        MoveToml,
-        NotRuntime,
-        Some("workers.queue-summary-secs"),
-        WarnEnvOverride,
-        false,
-    ),
-    spec(
-        "AXON_QDRANT_POINT_BUFFER",
-        MoveToml,
-        NotRuntime,
-        Some("workers.qdrant-point-buffer"),
-        WarnEnvOverride,
-        false,
-    ),
-    spec(
-        "AXON_MAX_PENDING_CRAWL_JOBS",
-        MoveToml,
-        NotRuntime,
-        Some("workers.max-pending-crawl-jobs"),
-        WarnEnvOverride,
-        false,
-    ),
-    spec(
-        "AXON_MAX_PENDING_EMBED_JOBS",
-        MoveToml,
-        NotRuntime,
-        Some("workers.max-pending-embed-jobs"),
-        WarnEnvOverride,
-        false,
-    ),
-    spec(
-        "AXON_MAX_PENDING_EXTRACT_JOBS",
-        MoveToml,
-        NotRuntime,
-        Some("workers.max-pending-extract-jobs"),
-        WarnEnvOverride,
-        false,
-    ),
-    spec(
-        "AXON_MAX_PENDING_INGEST_JOBS",
-        MoveToml,
-        NotRuntime,
-        Some("workers.max-pending-ingest-jobs"),
-        WarnEnvOverride,
-        false,
-    ),
-    spec(
-        "AXON_BIN",
-        TrustedOperatorBootstrap,
-        HostOnly,
-        None,
-        Advanced,
-        false,
-    ),
-    spec(
-        "AXON_CHROME_DIAGNOSTICS_DIR",
-        TrustedOperatorBootstrap,
-        HostOnly,
-        None,
-        Advanced,
-        false,
-    ),
-    spec(
-        "AXON_COLLECTION",
-        TrustedOperatorBootstrap,
-        HostOnly,
-        None,
-        Advanced,
-        false,
-    ),
-    spec(
-        "AXON_EXCLUDE_PATH_PREFIX",
-        TrustedOperatorBootstrap,
-        HostOnly,
-        None,
-        Advanced,
-        false,
-    ),
-    spec(
-        "AXON_INSTALL_PREFIX",
-        TrustedOperatorBootstrap,
-        HostOnly,
-        None,
-        Advanced,
-        false,
-    ),
-    spec(
-        "AXON_INSTALL_TMPDIR",
-        TrustedOperatorBootstrap,
-        HostOnly,
-        None,
-        Advanced,
-        false,
-    ),
-    spec(
-        "AXON_INSTALL_URL",
-        TrustedOperatorBootstrap,
-        HostOnly,
-        None,
-        Advanced,
-        false,
-    ),
-    spec(
-        "AXON_IN_CONTAINER",
-        ComposeEnv,
-        ContainerRequired,
-        None,
-        Advanced,
-        false,
-    ),
-    spec(
-        "AXON_LOG_DIR",
-        TrustedOperatorBootstrap,
-        HostOnly,
-        None,
-        Advanced,
-        false,
-    ),
-    spec(
-        "AXON_LOG_FILE",
-        TrustedOperatorBootstrap,
-        HostOnly,
-        None,
-        Advanced,
-        false,
-    ),
-    spec(
-        "AXON_LOG_MAX_FILES",
-        TrustedOperatorBootstrap,
-        HostOnly,
-        None,
-        Advanced,
-        false,
-    ),
-    spec(
-        "AXON_MCP_ARTIFACT_DIR",
-        TrustedOperatorBootstrap,
-        HostOnly,
-        None,
-        Advanced,
-        false,
-    ),
-    spec(
-        "AXON_MCP_EMBED_ALLOWED_ROOTS",
-        TrustedOperatorBootstrap,
-        HostOnly,
-        None,
-        Advanced,
-        false,
-    ),
-    spec(
-        "AXON_MCP_HTTP_HOST",
-        TrustedOperatorBootstrap,
-        HostOnly,
-        None,
-        Advanced,
-        false,
-    ),
-    spec(
-        "AXON_MCP_HTTP_PORT",
-        TrustedOperatorBootstrap,
-        HostOnly,
-        None,
-        Advanced,
-        false,
-    ),
-    spec(
-        "AXON_MCP_TRANSPORT",
-        TrustedOperatorBootstrap,
-        HostOnly,
-        None,
-        Advanced,
-        false,
-    ),
-    spec(
-        "AXON_NEO4J_URL",
-        TrustedOperatorBootstrap,
-        HostOnly,
-        None,
-        Advanced,
-        false,
-    ),
-    spec(
-        "AXON_OUTPUT_DIR",
-        TrustedOperatorBootstrap,
-        HostOnly,
-        None,
-        Advanced,
-        false,
-    ),
-    spec(
-        "AXON_REPO_ROOT",
-        TrustedOperatorBootstrap,
-        HostOnly,
-        None,
-        Advanced,
-        false,
-    ),
-    spec(
-        "AXON_SUGGEST_BASE_URL_LIMIT",
-        TrustedOperatorBootstrap,
-        HostOnly,
-        None,
-        Advanced,
-        false,
-    ),
-    spec(
-        "AXON_SUGGEST_EXISTING_URL_LIMIT",
-        TrustedOperatorBootstrap,
-        HostOnly,
-        None,
-        Advanced,
-        false,
-    ),
-    spec(
-        "AXON_TARGET_DIR",
-        TrustedOperatorBootstrap,
-        HostOnly,
-        None,
-        Advanced,
-        false,
-    ),
-    spec(
-        "AXON_WEB_ALLOWED_ORIGINS",
-        TrustedOperatorBootstrap,
-        HostOnly,
-        None,
-        Advanced,
-        false,
-    ),
-    spec(
-        "CUDA_CACHE_DISABLE",
-        ComposeEnv,
-        ComposeInterpolation,
-        None,
-        Advanced,
-        false,
-    ),
-    spec(
-        "GEMINI_AUTH_FILES",
-        TrustedOperatorBootstrap,
-        HostOnly,
-        None,
-        Advanced,
-        false,
-    ),
-    spec(
-        "GOOGLE_APPLICATION_CREDENTIALS",
-        TrustedOperatorBootstrap,
-        Both,
-        None,
-        Advanced,
-        false,
-    ),
-    spec(
-        "HF_HUB_CACHE",
-        ComposeEnv,
-        ComposeInterpolation,
-        None,
-        Advanced,
-        false,
-    ),
-    spec(
-        "HF_HUB_ENABLE_HF_TRANSFER",
-        ComposeEnv,
-        ComposeInterpolation,
-        None,
-        Advanced,
-        false,
-    ),
-    spec(
-        "NVIDIA_REQUIRE_CUDA",
-        ComposeEnv,
-        ComposeInterpolation,
-        None,
-        Advanced,
-        false,
-    ),
-    spec(
-        "QDRANT_COLLECTION",
-        TrustedOperatorBootstrap,
-        HostOnly,
-        None,
-        Advanced,
-        false,
-    ),
-    spec(
-        "TEI_MAX_BATCH_REQUESTS",
-        ComposeEnv,
-        ComposeInterpolation,
-        None,
-        Advanced,
-        false,
-    ),
-    spec(
-        "TEI_MAX_CONCURRENT_REQUESTS",
-        ComposeEnv,
-        ComposeInterpolation,
-        None,
-        Advanced,
-        false,
-    ),
-    spec(
-        "TEI_POOLING",
-        ComposeEnv,
-        ComposeInterpolation,
-        None,
-        Advanced,
-        false,
-    ),
-    spec(
-        "AXON_ASK_CANDIDATE_LIMIT",
-        MoveToml,
-        NotRuntime,
-        Some("ask.candidate-limit"),
-        WarnEnvOverride,
-        false,
-    ),
-    spec(
-        "AXON_ASK_CHUNK_LIMIT",
-        MoveToml,
-        NotRuntime,
-        Some("ask.chunk-limit"),
-        WarnEnvOverride,
-        false,
-    ),
-    spec(
-        "AXON_ASK_MIN_RELEVANCE_SCORE",
-        MoveToml,
-        NotRuntime,
-        Some("ask.min-relevance-score"),
-        WarnEnvOverride,
-        false,
-    ),
-    spec(
-        "AXON_ASK_HYBRID_CANDIDATES",
-        MoveToml,
-        NotRuntime,
-        Some("search.ask-hybrid-candidates"),
-        WarnEnvOverride,
-        false,
-    ),
-    spec(
-        "AXON_HYBRID_SEARCH",
-        MoveToml,
-        NotRuntime,
-        Some("search.hybrid-enabled"),
-        WarnEnvOverride,
-        false,
-    ),
-    spec(
-        "AXON_HYBRID_CANDIDATES",
-        MoveToml,
-        NotRuntime,
-        Some("search.hybrid-candidates"),
-        WarnEnvOverride,
-        false,
-    ),
-    spec(
-        "AXON_HNSW_EF_SEARCH",
-        MoveToml,
-        NotRuntime,
-        Some("search.hnsw-ef"),
-        WarnEnvOverride,
-        false,
-    ),
-    spec(
-        "AXON_HNSW_EF_SEARCH_LEGACY",
-        MoveToml,
-        NotRuntime,
-        Some("search.hnsw-ef-legacy"),
-        WarnEnvOverride,
-        false,
-    ),
-    spec(
-        "AXON_BATCH_QUEUE",
-        Delete,
-        NotRuntime,
-        None,
-        DeleteOnMigration,
-        false,
-    ),
-    spec(
-        "AXON_CRAWL_QUEUE",
-        Delete,
-        NotRuntime,
-        None,
-        DeleteOnMigration,
-        false,
-    ),
-    spec(
-        "AXON_EMBED_QUEUE",
-        Delete,
-        NotRuntime,
-        None,
-        DeleteOnMigration,
-        false,
-    ),
-    spec(
-        "AXON_EXTRACT_QUEUE",
-        Delete,
-        NotRuntime,
-        None,
-        DeleteOnMigration,
-        false,
-    ),
-    spec(
-        "AXON_INGEST_QUEUE",
-        Delete,
-        NotRuntime,
-        None,
-        DeleteOnMigration,
-        false,
-    ),
-    spec(
-        "AXON_AMQP_URL",
-        Delete,
-        NotRuntime,
-        None,
-        DeleteOnMigration,
-        false,
-    ),
-    spec(
-        "AXON_LITE",
-        Delete,
-        NotRuntime,
-        None,
-        DeleteOnMigration,
-        false,
-    ),
-    spec(
-        "AXON_PG_MCP_URL",
-        Delete,
-        NotRuntime,
-        None,
-        DeleteOnMigration,
-        false,
-    ),
-    spec(
-        "AXON_PG_URL",
-        Delete,
-        NotRuntime,
-        None,
-        DeleteOnMigration,
-        false,
-    ),
-    spec(
-        "AXON_REDIS_URL",
-        Delete,
-        NotRuntime,
-        None,
-        DeleteOnMigration,
-        false,
-    ),
-];
+#[path = "env_registry/advanced.rs"]
+mod advanced;
+#[path = "env_registry/migration.rs"]
+mod migration;
+#[path = "env_registry/runtime.rs"]
+mod runtime;
 
-const fn spec(
+pub(crate) const fn spec(
     key: &'static str,
     classification: EnvClassification,
     placement: RuntimePlacement,
@@ -811,13 +62,24 @@ const fn spec(
     }
 }
 
+pub(crate) fn all_specs() -> impl Iterator<Item = &'static EnvKeySpec> {
+    [
+        runtime::RUNTIME_ENV_KEY_SPECS,
+        advanced::ADVANCED_ENV_KEY_SPECS,
+        migration::MIGRATION_ENV_KEY_SPECS,
+    ]
+    .into_iter()
+    .flatten()
+}
+
 pub(crate) fn spec_for(key: &str) -> Option<&'static EnvKeySpec> {
-    ENV_KEY_SPECS.iter().find(|spec| spec.key == key)
+    all_specs().find(|spec| spec.key == key)
 }
 
 #[cfg(test)]
 mod tests {
     use super::*;
+    use EnvClassification::{KeepEnv, MoveToml};
 
     #[test]
     fn service_urls_are_env_not_toml() {
@@ -830,7 +92,7 @@ mod tests {
 
     #[test]
     fn moved_tuning_has_toml_destination() {
-        for spec in ENV_KEY_SPECS {
+        for spec in all_specs() {
             if spec.classification == MoveToml {
                 assert!(
                     spec.toml_destination.is_some(),
