+++ src/core/config/parse/env_registry/advanced.rs
@@ -1,0 +1,383 @@
+use super::EnvClassification::*;
+use super::LegacyBehavior::*;
+use super::RuntimePlacement::*;
+use super::{EnvKeySpec, spec};
+
+pub(crate) const ADVANCED_ENV_KEY_SPECS: &[EnvKeySpec] = &[
+    spec(
+        "AXON_ENV_FILE",
+        TrustedOperatorBootstrap,
+        HostOnly,
+        None,
+        Advanced,
+        false,
+    ),
+    spec(
+        "AXON_CONFIG_PATH",
+        TrustedOperatorBootstrap,
+        HostOnly,
+        None,
+        Advanced,
+        false,
+    ),
+    spec(
+        "AXON_HOME",
+        TrustedOperatorBootstrap,
+        Both,
+        None,
+        Advanced,
+        false,
+    ),
+    spec(
+        "AXON_DATA_DIR",
+        TrustedOperatorBootstrap,
+        Both,
+        None,
+        Advanced,
+        false,
+    ),
+    spec(
+        "AXON_SQLITE_PATH",
+        TrustedOperatorBootstrap,
+        Both,
+        None,
+        Advanced,
+        false,
+    ),
+    spec(
+        "AXON_MCP_HTTP_PUBLISH",
+        ComposeEnv,
+        ComposeInterpolation,
+        None,
+        Canonical,
+        false,
+    ),
+    spec(
+        "AXON_IMAGE",
+        ComposeEnv,
+        ComposeInterpolation,
+        None,
+        Canonical,
+        false,
+    ),
+    spec(
+        "GEMINI_HOME",
+        ComposeEnv,
+        ComposeInterpolation,
+        None,
+        Canonical,
+        false,
+    ),
+    spec(
+        "TEI_EMBEDDING_MODEL",
+        ComposeEnv,
+        ComposeInterpolation,
+        None,
+        Canonical,
+        false,
+    ),
+    spec(
+        "TEI_HTTP_PORT",
+        ComposeEnv,
+        ComposeInterpolation,
+        None,
+        Canonical,
+        false,
+    ),
+    spec(
+        "TEI_SERVER_MAX_CLIENT_BATCH_SIZE",
+        ComposeEnv,
+        ComposeInterpolation,
+        None,
+        Canonical,
+        false,
+    ),
+    spec(
+        "NVIDIA_VISIBLE_DEVICES",
+        ComposeEnv,
+        ComposeInterpolation,
+        None,
+        Canonical,
+        false,
+    ),
+    spec(
+        "CUDA_VISIBLE_DEVICES",
+        ComposeEnv,
+        ComposeInterpolation,
+        None,
+        Canonical,
+        false,
+    ),
+    spec(
+        "AXON_HEADLESS_GEMINI_HOME",
+        TrustedOperatorBootstrap,
+        Both,
+        None,
+        Advanced,
+        false,
+    ),
+    spec(
+        "AXON_BIN",
+        TrustedOperatorBootstrap,
+        HostOnly,
+        None,
+        Advanced,
+        false,
+    ),
+    spec(
+        "AXON_CHROME_DIAGNOSTICS_DIR",
+        TrustedOperatorBootstrap,
+        HostOnly,
+        None,
+        Advanced,
+        false,
+    ),
+    spec(
+        "AXON_COLLECTION",
+        TrustedOperatorBootstrap,
+        HostOnly,
+        None,
+        Advanced,
+        false,
+    ),
+    spec(
+        "AXON_EXCLUDE_PATH_PREFIX",
+        TrustedOperatorBootstrap,
+        HostOnly,
+        None,
+        Advanced,
+        false,
+    ),
+    spec(
+        "AXON_INSTALL_PREFIX",
+        TrustedOperatorBootstrap,
+        HostOnly,
+        None,
+        Advanced,
+        false,
+    ),
+    spec(
+        "AXON_INSTALL_TMPDIR",
+        TrustedOperatorBootstrap,
+        HostOnly,
+        None,
+        Advanced,
+        false,
+    ),
+    spec(
+        "AXON_INSTALL_URL",
+        TrustedOperatorBootstrap,
+        HostOnly,
+        None,
+        Advanced,
+        false,
+    ),
+    spec(
+        "AXON_IN_CONTAINER",
+        ComposeEnv,
+        ContainerRequired,
+        None,
+        Advanced,
+        false,
+    ),
+    spec(
+        "AXON_LOG_DIR",
+        TrustedOperatorBootstrap,
+        HostOnly,
+        None,
+        Advanced,
+        false,
+    ),
+    spec(
+        "AXON_LOG_FILE",
+        TrustedOperatorBootstrap,
+        HostOnly,
+        None,
+        Advanced,
+        false,
+    ),
+    spec(
+        "AXON_LOG_MAX_FILES",
+        TrustedOperatorBootstrap,
+        HostOnly,
+        None,
+        Advanced,
+        false,
+    ),
+    spec(
+        "AXON_MCP_ARTIFACT_DIR",
+        TrustedOperatorBootstrap,
+        HostOnly,
+        None,
+        Advanced,
+        false,
+    ),
+    spec(
+        "AXON_MCP_EMBED_ALLOWED_ROOTS",
+        TrustedOperatorBootstrap,
+        HostOnly,
+        None,
+        Advanced,
+        false,
+    ),
+    spec(
+        "AXON_MCP_HTTP_HOST",
+        TrustedOperatorBootstrap,
+        HostOnly,
+        None,
+        Advanced,
+        false,
+    ),
+    spec(
+        "AXON_MCP_HTTP_PORT",
+        TrustedOperatorBootstrap,
+        HostOnly,
+        None,
+        Advanced,
+        false,
+    ),
+    spec(
+        "AXON_MCP_TRANSPORT",
+        TrustedOperatorBootstrap,
+        HostOnly,
+        None,
+        Advanced,
+        false,
+    ),
+    spec(
+        "AXON_NEO4J_URL",
+        TrustedOperatorBootstrap,
+        HostOnly,
+        None,
+        Advanced,
+        false,
+    ),
+    spec(
+        "AXON_OUTPUT_DIR",
+        TrustedOperatorBootstrap,
+        HostOnly,
+        None,
+        Advanced,
+        false,
+    ),
+    spec(
+        "AXON_REPO_ROOT",
+        TrustedOperatorBootstrap,
+        HostOnly,
+        None,
+        Advanced,
+        false,
+    ),
+    spec(
+        "AXON_SUGGEST_BASE_URL_LIMIT",
+        TrustedOperatorBootstrap,
+        HostOnly,
+        None,
+        Advanced,
+        false,
+    ),
+    spec(
+        "AXON_SUGGEST_EXISTING_URL_LIMIT",
+        TrustedOperatorBootstrap,
+        HostOnly,
+        None,
+        Advanced,
+        false,
+    ),
+    spec(
+        "AXON_TARGET_DIR",
+        TrustedOperatorBootstrap,
+        HostOnly,
+        None,
+        Advanced,
+        false,
+    ),
+    spec(
+        "AXON_WEB_ALLOWED_ORIGINS",
+        TrustedOperatorBootstrap,
+        HostOnly,
+        None,
+        Advanced,
+        false,
+    ),
+    spec(
+        "CUDA_CACHE_DISABLE",
+        ComposeEnv,
+        ComposeInterpolation,
+        None,
+        Advanced,
+        false,
+    ),
+    spec(
+        "GEMINI_AUTH_FILES",
+        TrustedOperatorBootstrap,
+        HostOnly,
+        None,
+        Advanced,
+        false,
+    ),
+    spec(
+        "GOOGLE_APPLICATION_CREDENTIALS",
+        TrustedOperatorBootstrap,
+        Both,
+        None,
+        Advanced,
+        false,
+    ),
+    spec(
+        "HF_HUB_CACHE",
+        ComposeEnv,
+        ComposeInterpolation,
+        None,
+        Advanced,
+        false,
+    ),
+    spec(
+        "HF_HUB_ENABLE_HF_TRANSFER",
+        ComposeEnv,
+        ComposeInterpolation,
+        None,
+        Advanced,
+        false,
+    ),
+    spec(
+        "NVIDIA_REQUIRE_CUDA",
+        ComposeEnv,
+        ComposeInterpolation,
+        None,
+        Advanced,
+        false,
+    ),
+    spec(
+        "QDRANT_COLLECTION",
+        TrustedOperatorBootstrap,
+        HostOnly,
+        None,
+        Advanced,
+        false,
+    ),
+    spec(
+        "TEI_MAX_BATCH_REQUESTS",
+        ComposeEnv,
+        ComposeInterpolation,
+        None,
+        Advanced,
+        false,
+    ),
+    spec(
+        "TEI_MAX_CONCURRENT_REQUESTS",
+        ComposeEnv,
+        ComposeInterpolation,
+        None,
+        Advanced,
+        false,
+    ),
+    spec(
+        "TEI_POOLING",
+        ComposeEnv,
+        ComposeInterpolation,
+        None,
+        Advanced,
+        false,
+    ),
+];
