+++ src/core/config/parse/env_registry/migration.rs
@@ -1,0 +1,247 @@
+use super::EnvClassification::*;
+use super::LegacyBehavior::*;
+use super::RuntimePlacement::*;
+use super::{EnvKeySpec, spec};
+
+pub(crate) const MIGRATION_ENV_KEY_SPECS: &[EnvKeySpec] = &[
+    spec(
+        "TEI_MAX_CLIENT_BATCH_SIZE",
+        MoveToml,
+        NotRuntime,
+        Some("tei.max-client-batch-size"),
+        WarnEnvOverride,
+        false,
+    ),
+    spec(
+        "TEI_MAX_RETRIES",
+        MoveToml,
+        NotRuntime,
+        Some("tei.max-retries"),
+        WarnEnvOverride,
+        false,
+    ),
+    spec(
+        "TEI_REQUEST_TIMEOUT_MS",
+        MoveToml,
+        NotRuntime,
+        Some("tei.request-timeout-ms"),
+        WarnEnvOverride,
+        false,
+    ),
+    spec(
+        "AXON_INGEST_LANES",
+        MoveToml,
+        NotRuntime,
+        Some("workers.ingest-lanes"),
+        WarnEnvOverride,
+        false,
+    ),
+    spec(
+        "AXON_EMBED_LANES",
+        MoveToml,
+        NotRuntime,
+        Some("workers.embed-lanes"),
+        WarnEnvOverride,
+        false,
+    ),
+    spec(
+        "AXON_EMBED_DOC_TIMEOUT_SECS",
+        MoveToml,
+        NotRuntime,
+        Some("workers.embed-doc-timeout-secs"),
+        WarnEnvOverride,
+        false,
+    ),
+    spec(
+        "AXON_QUEUE_SUMMARY_SECS",
+        MoveToml,
+        NotRuntime,
+        Some("workers.queue-summary-secs"),
+        WarnEnvOverride,
+        false,
+    ),
+    spec(
+        "AXON_QDRANT_POINT_BUFFER",
+        MoveToml,
+        NotRuntime,
+        Some("workers.qdrant-point-buffer"),
+        WarnEnvOverride,
+        false,
+    ),
+    spec(
+        "AXON_MAX_PENDING_CRAWL_JOBS",
+        MoveToml,
+        NotRuntime,
+        Some("workers.max-pending-crawl-jobs"),
+        WarnEnvOverride,
+        false,
+    ),
+    spec(
+        "AXON_MAX_PENDING_EMBED_JOBS",
+        MoveToml,
+        NotRuntime,
+        Some("workers.max-pending-embed-jobs"),
+        WarnEnvOverride,
+        false,
+    ),
+    spec(
+        "AXON_MAX_PENDING_EXTRACT_JOBS",
+        MoveToml,
+        NotRuntime,
+        Some("workers.max-pending-extract-jobs"),
+        WarnEnvOverride,
+        false,
+    ),
+    spec(
+        "AXON_MAX_PENDING_INGEST_JOBS",
+        MoveToml,
+        NotRuntime,
+        Some("workers.max-pending-ingest-jobs"),
+        WarnEnvOverride,
+        false,
+    ),
+    spec(
+        "AXON_ASK_CANDIDATE_LIMIT",
+        MoveToml,
+        NotRuntime,
+        Some("ask.candidate-limit"),
+        WarnEnvOverride,
+        false,
+    ),
+    spec(
+        "AXON_ASK_CHUNK_LIMIT",
+        MoveToml,
+        NotRuntime,
+        Some("ask.chunk-limit"),
+        WarnEnvOverride,
+        false,
+    ),
+    spec(
+        "AXON_ASK_MIN_RELEVANCE_SCORE",
+        MoveToml,
+        NotRuntime,
+        Some("ask.min-relevance-score"),
+        WarnEnvOverride,
+        false,
+    ),
+    spec(
+        "AXON_ASK_HYBRID_CANDIDATES",
+        MoveToml,
+        NotRuntime,
+        Some("search.ask-hybrid-candidates"),
+        WarnEnvOverride,
+        false,
+    ),
+    spec(
+        "AXON_HYBRID_SEARCH",
+        MoveToml,
+        NotRuntime,
+        Some("search.hybrid-enabled"),
+        WarnEnvOverride,
+        false,
+    ),
+    spec(
+        "AXON_HYBRID_CANDIDATES",
+        MoveToml,
+        NotRuntime,
+        Some("search.hybrid-candidates"),
+        WarnEnvOverride,
+        false,
+    ),
+    spec(
+        "AXON_HNSW_EF_SEARCH",
+        MoveToml,
+        NotRuntime,
+        Some("search.hnsw-ef"),
+        WarnEnvOverride,
+        false,
+    ),
+    spec(
+        "AXON_HNSW_EF_SEARCH_LEGACY",
+        MoveToml,
+        NotRuntime,
+        Some("search.hnsw-ef-legacy"),
+        WarnEnvOverride,
+        false,
+    ),
+    spec(
+        "AXON_BATCH_QUEUE",
+        Delete,
+        NotRuntime,
+        None,
+        DeleteOnMigration,
+        false,
+    ),
+    spec(
+        "AXON_CRAWL_QUEUE",
+        Delete,
+        NotRuntime,
+        None,
+        DeleteOnMigration,
+        false,
+    ),
+    spec(
+        "AXON_EMBED_QUEUE",
+        Delete,
+        NotRuntime,
+        None,
+        DeleteOnMigration,
+        false,
+    ),
+    spec(
+        "AXON_EXTRACT_QUEUE",
+        Delete,
+        NotRuntime,
+        None,
+        DeleteOnMigration,
+        false,
+    ),
+    spec(
+        "AXON_INGEST_QUEUE",
+        Delete,
+        NotRuntime,
+        None,
+        DeleteOnMigration,
+        false,
+    ),
+    spec(
+        "AXON_AMQP_URL",
+        Delete,
+        NotRuntime,
+        None,
+        DeleteOnMigration,
+        false,
+    ),
+    spec(
+        "AXON_LITE",
+        Delete,
+        NotRuntime,
+        None,
+        DeleteOnMigration,
+        false,
+    ),
+    spec(
+        "AXON_PG_MCP_URL",
+        Delete,
+        NotRuntime,
+        None,
+        DeleteOnMigration,
+        false,
+    ),
+    spec(
+        "AXON_PG_URL",
+        Delete,
+        NotRuntime,
+        None,
+        DeleteOnMigration,
+        false,
+    ),
+    spec(
+        "AXON_REDIS_URL",
+        Delete,
+        NotRuntime,
+        None,
+        DeleteOnMigration,
+        false,
+    ),
+];
