+++ src/core/config/parse/env_registry/runtime.rs
@@ -1,0 +1,136 @@
+use super::EnvClassification::*;
+use super::LegacyBehavior::*;
+use super::RuntimePlacement::*;
+use super::{EnvKeySpec, spec};
+
+pub(crate) const RUNTIME_ENV_KEY_SPECS: &[EnvKeySpec] = &[
+    spec("QDRANT_URL", KeepEnv, Both, None, Canonical, false),
+    spec("TEI_URL", KeepEnv, Both, None, Canonical, false),
+    spec(
+        "AXON_CHROME_REMOTE_URL",
+        KeepEnv,
+        Both,
+        None,
+        Canonical,
+        false,
+    ),
+    spec("AXON_SERVER_URL", KeepEnv, HostOnly, None, Canonical, false),
+    spec(
+        "AXON_MCP_HTTP_TOKEN",
+        KeepEnv,
+        ContainerRequired,
+        None,
+        Canonical,
+        true,
+    ),
+    spec("AXON_MCP_AUTH_MODE", KeepEnv, Both, None, Canonical, false),
+    spec("AXON_MCP_PUBLIC_URL", KeepEnv, Both, None, Canonical, false),
+    spec(
+        "AXON_MCP_GOOGLE_CLIENT_ID",
+        KeepEnv,
+        Both,
+        None,
+        Canonical,
+        false,
+    ),
+    spec(
+        "AXON_MCP_GOOGLE_CLIENT_SECRET",
+        KeepEnv,
+        Both,
+        None,
+        Canonical,
+        true,
+    ),
+    spec(
+        "AXON_MCP_AUTH_ADMIN_EMAIL",
+        KeepEnv,
+        Both,
+        None,
+        Canonical,
+        false,
+    ),
+    spec(
+        "AXON_MCP_AUTH_ALLOWED_REDIRECT_URIS",
+        KeepEnv,
+        Both,
+        None,
+        Canonical,
+        false,
+    ),
+    spec(
+        "AXON_MCP_ALLOWED_ORIGINS",
+        KeepEnv,
+        Both,
+        None,
+        Canonical,
+        false,
+    ),
+    spec("TAVILY_API_KEY", KeepEnv, Both, None, Canonical, true),
+    spec("GITHUB_TOKEN", KeepEnv, Both, None, Canonical, true),
+    spec("REDDIT_CLIENT_ID", KeepEnv, Both, None, Canonical, false),
+    spec("REDDIT_CLIENT_SECRET", KeepEnv, Both, None, Canonical, true),
+    spec(
+        "HF_TOKEN",
+        KeepEnv,
+        ComposeInterpolation,
+        None,
+        Canonical,
+        true,
+    ),
+    spec(
+        "OPENAI_MODEL",
+        CompatibilityShim,
+        Both,
+        None,
+        WarnEnvOverride,
+        false,
+    ),
+    spec(
+        "OPENAI_BASE_URL",
+        CompatibilityShim,
+        Both,
+        None,
+        WarnAndIgnore,
+        false,
+    ),
+    spec(
+        "OPENAI_API_KEY",
+        CompatibilityShim,
+        Both,
+        None,
+        WarnAndIgnore,
+        true,
+    ),
+    spec(
+        "AXON_HEADLESS_GEMINI_CMD",
+        KeepEnv,
+        Both,
+        None,
+        Canonical,
+        false,
+    ),
+    spec(
+        "AXON_HEADLESS_GEMINI_MODEL",
+        KeepEnv,
+        Both,
+        None,
+        Canonical,
+        false,
+    ),
+    spec(
+        "AXON_LLM_COMPLETION_CONCURRENCY",
+        KeepEnv,
+        Both,
+        None,
+        Canonical,
+        false,
+    ),
+    spec(
+        "AXON_LLM_COMPLETION_TIMEOUT_SECS",
+        KeepEnv,
+        Both,
+        None,
+        Canonical,
+        false,
+    ),
+];
