+++ src/core/config/types/config_impls.rs
@@ -338,7 +338,10 @@ impl fmt::Debug for Config {
             .field("tei_request_timeout_ms", &self.tei_request_timeout_ms)
             .field("tei_max_client_batch_size", &self.tei_max_client_batch_size)
             .field("ingest_lanes", &self.ingest_lanes)
+            .field("embed_lanes", &self.embed_lanes)
             .field("embed_doc_timeout_secs", &self.embed_doc_timeout_secs)
+            .field("queue_summary_secs", &self.queue_summary_secs)
+            .field("qdrant_point_buffer", &self.qdrant_point_buffer)
             .field("max_pending_crawl_jobs", &self.max_pending_crawl_jobs)
             .field("max_pending_embed_jobs", &self.max_pending_embed_jobs)
             .field("max_pending_extract_jobs", &self.max_pending_extract_jobs)
